feat: expose `getTags` as a public method
Basically, in Workers, there are some really
unsual situation where ALS (AsyncLocalStorage) doesn't
work - where context is lost.

For instance, it happens when you cross JSRPC
boundaries twice:

DO (caller) <-> Worker (calee) <-> RpcTarget (caller)

RpcTarget in this case will not have the AsyncContext propagated, so we
need to inject it in a different way. This PR allows to expose all tags
so that we can inject the logger using a different method.