chore: remove Zod dependency from workers-tagged-logger (#72)
workers-tagged-logger no longer requires Zod as a peer dependency.
Zod was only used for generating types used for testing, and should not
have been exported.

Consumers of this package shouldn't be relying on these Zod schemas, so
hopefully this is not a breaking change for anyone.

* chore: remove unused build tests command

* chore: remove workers-tagged-logger integration tests

No longer need to test against zod 3 / 4 because Zod is no longer a
dependency of workers-tagged-logger

* chore: remove install-deps script (no longer needed)

* chore: remove Zod schema imports in tests

workers-tagged-logger no longer exports Zod schemas, so we had to change
to type imports in these tests

* fix: switch to type exports

* fix: remove tests dir from Earthfile (no longer exists)

* fix: don't run nested turbo command

I think this is causing race conditions leading to build failures

* chore: update pnpm

* fix: remove inject-workspace-packages (was causing issues)